requirements.txt
streamlit
pandas
streamlit-aggrid
python-docx
Pillow